+++ packages/components/src/index.ts
@@ -1,3 +1,5 @@
+export * from "./Button"
+export * from "./ClearButton"
 export * from "./Filter"
 export * from "./FilterBar"
 export * from "./FilterButton"
@@ -5,6 +7,7 @@ export * from "./FilterDatePicker"
 export * from "./FilterDateRangePicker"
 export * from "./FilterMultiSelect"
 export * from "./FilterSelect"
+export * from "./InputSearch"
 export * from "./KaizenProvider"
 export * from "./Workflow"